requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
PyMuPDF>=1.23.0
